% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn.R
\name{espn_standings}
\alias{espn_standings}
\title{Get team standings for a sport/league}
\usage{
espn_standings(sport = "football", league = "nfl")
}
\arguments{
\item{sport}{Sport: "football", "basketball", "baseball", "hockey", "soccer"}

\item{league}{League: "nfl", "nba", "mlb", "nhl", "eng.1"}
}
\value{
tibble: team, wins, losses, pct, streak, division
}
\description{
Get team standings for a sport/league
}
