% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn.R
\name{espn_teams}
\alias{espn_teams}
\title{Get teams for a sport/league}
\usage{
espn_teams(sport = "football", league = "nfl")
}
\arguments{
\item{sport}{Sport: "football", "basketball", "baseball", "hockey", "soccer"}

\item{league}{League: "nfl", "nba", "mlb", "nhl", "eng.1"}
}
\value{
tibble: id, name, abbreviation, display_name, location, logo
}
\description{
Get teams for a sport/league
}
