% Generated by build script
\name{espn_scores}
\alias{espn_scores}
\title{Get ESPN scores for a sport}
\usage{
espn_scores()
}
\arguments{
\item{sport}{Sport ("football", "basketball", "baseball", "hockey")}
\item{league}{League ("nfl", "nba", "mlb", "nhl")}
}
\value{tibble of recent scores}
\description{Get ESPN scores for a sport}
