% Generated by build script
\name{espn_standings}
\alias{espn_standings}
\title{Get ESPN standings}
\usage{
espn_standings()
}
\arguments{
\item{sport}{Sport ("football", "basketball", "baseball", "hockey")}
\item{league}{League ("nfl", "nba", "mlb", "nhl")}
}
\value{tibble of standings}
\description{Get ESPN standings}
